+++ packages/css-packs/reduced-ui-pack/examples/02-button.js
@@ -1,11 +1,10 @@
 // @flow
 import React from 'react';
 // eslint-disable-next-line
-import reducedStyles from '!!raw-loader!../src/bundle.css';
+import stuff from '!!style-loader!css-loader!../src/bundle.css';
 
 export default () => (
   <div>
-    <style>{reducedStyles}</style>
     <div title="Buttons">
       <h1>Buttons</h1>
       <h2>default appearance</h2>