+++ packages/editor/editor-common/src/utils/validator.ts
@@ -315,7 +315,7 @@ export const getValidNode = (
           details &&
           details.some(meta => {
             const { badge, lozenge, users } = meta;
-            if (badge && !badge.value) {
+            if (badge && typeof badge.value !== 'number') {
               return true;
             }
             if (lozenge && !lozenge.text) {
@@ -698,6 +698,13 @@ export const getValidNode = (
             attrs,
           };
         }
+        break;
+      }
+      case 'placeholder': {
+        if (attrs && typeof attrs.text !== 'undefined') {
+          return { type, attrs };
+        }
+
         break;
       }
     }