+++ packages/editor/editor-core/src/ui/Appearance/FullPage.tsx
@@ -11,7 +11,7 @@ import ContentStyles from '../ContentStyles';
 import { ClickAreaBlock } from '../Addon';
 import WidthDetector from '../WidthDetector';
 
-const GUTTER_PADDING = 26;
+const GUTTER_PADDING = 32;
 
 const FullPageEditorWrapper = styled.div`
   min-width: 340px;
@@ -32,6 +32,7 @@ const ScrollContainer = styled(ContentStyles)`
 ScrollContainer.displayName = 'ScrollContainer';
 
 const ContentArea = styled.div`
+  line-height: 24px;
   height: 100%;
   width: 100%;
   max-width: ${akEditorFullPageMaxWidth + GUTTER_PADDING * 2}px;